[DM-21332] Turn on NFSv4 for fileserver
This allows us to run a fileserver that supports NFSv4.

The exports file is particularly important.  The first line, with
fsid=0 is important, since it is the root of what is exported.
If this isn't rw, then nothing under it can be rw.  If this isn't
no_root_squash, then everything under it can't be no_root_squash,
even if explicitly saying otherwise.  sec=sys tells NFS to not
try to map the account names to their UIDs (which NFSv4 supports,
to allow for the same user with different UIDs on the server
and client to work seamlessly).  Since the fileserver doesn't
have these accounts, this won't work.  We just want to trust the
uid like NFSv3, which is done by sec=sys.  This also must be on
the /exports otherwise it doesn't seem to be respected.

When mounting this on NFSv4, you would mount /home, not /exports/home,
which is different than v3, just to make it a little more confusing.